requirements.txt
apache-airflow==2.10.5
Authlib==1.4.1
cassandra_driver==3.29.2
category_encoders==2.6.3
fastapi==0.115.8
matplotlib==3.10.0
mysql_connector_repackaged==0.3.1
numpy==2.2.2
pandas==2.2.3
plotly==5.24.1
polars==1.22.0
psycopg2==2.9.10
pyarrow==17.0.0
pymongo==4.11
scikit_learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
starlette==0.45.3
uvicorn==0.34.0